Move data display objects into separate folder
