+++ src/main/java/controllers/data_objects/StatsDisplayData.java
@@ -1,4 +1,4 @@
-package ui.windows.stats;
+package controllers.data_objects;
 
 /**
  * Stores data to be displayed by a StatsDisplay.
